% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{assign_value}
\alias{assign_value}
\title{Action to assign value in selected cell}
\usage{
assign_value(mat, row, col, val)
}
\arguments{
\item{mat}{sudoku class object}

\item{row}{row index of sudoku matrix}

\item{col}{column index of sudoku matrix}

\item{val}{value to be filled}
}
\value{
modified sudoku class object
}
\description{
Action to assign value in selected cell
}
