% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_cols}
\alias{check_cols}
\title{Check each column agains predefined criteria}
\usage{
check_cols(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
vector with erroneus columns
}
\description{
Check each column agains predefined criteria
}
