% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{check_if_win}
\alias{check_if_win}
\title{Function to check if user wins}
\usage{
check_if_win(mat)
}
\arguments{
\item{mat}{sudoku class object}
}
\value{
boolean flag if win
}
\description{
Function to check if user wins
}
