% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_matrix}
\alias{check_matrix}
\title{Main function to check if sudoku matrix is correct}
\usage{
check_matrix(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
list where first element is global flag indicating if matrix is correct,
second element with erroneus rows,
third element with erroneus columns,
fourth element with erroneus squares
}
\description{
Main function to check if sudoku matrix is correct
}
