% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_range}
\alias{check_range}
\title{Check condition of values being within 1-9 range}
\usage{
check_range(values)
}
\arguments{
\item{values}{vector of values to be checked}
}
\value{
boolan flag
}
\description{
Check condition of values being within 1-9 range
}
