% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_rows}
\alias{check_rows}
\title{Check each row agains predefined criteria}
\usage{
check_rows(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
vector with erroneus rows
}
\description{
Check each row agains predefined criteria
}
