% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_squares}
\alias{check_squares}
\title{Check each mini-square agains predefined criteria}
\usage{
check_squares(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
list with erroneus squares in form of big indexes
}
\description{
Check each mini-square agains predefined criteria
}
