% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_unique}
\alias{check_unique}
\title{Check condition of values being unique}
\usage{
check_unique(values)
}
\arguments{
\item{values}{vector of values to be checked}
}
\value{
boolan flag
}
\description{
Check condition of values being unique
}
