% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{fill_NA}
\alias{fill_NA}
\title{Randomly fill empty values in matrix depending on difficulty}
\usage{
fill_NA(mat, difficulty)
}
\arguments{
\item{mat}{matrix or sudoku class object}

\item{difficulty}{game difficulty level. Three levels allowed (E)asy, (M)edium, (H)ard}
}
\description{
Randomly fill empty values in matrix depending on difficulty
}
