% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{flip}
\alias{flip}
\title{Flip matrix horizontally or vertically}
\usage{
flip(mat, if_row = TRUE)
}
\arguments{
\item{mat}{matrix or sudoku class object}

\item{if_row}{boolean flag indicating if horizontal or vertical flip}
}
\value{
modified matrix
}
\description{
Flip matrix horizontally or vertically
}
