% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{generate_matrix}
\alias{generate_matrix}
\title{Main function to generate final sudoku matrix}
\usage{
generate_matrix(difficulty)
}
\arguments{
\item{difficulty}{game difficulty level. Three levels allowed (E)asy, (M)edium, (H)ard}
}
\description{
Main function to generate final sudoku matrix
}
