% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{generate_raw_matrix}
\alias{generate_raw_matrix}
\title{Generate correct sudoku matrix to be modified}
\usage{
generate_raw_matrix()
}
\value{
correct sudoku matrix
}
\description{
Generate correct sudoku matrix to be modified
}
