% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{get_small_index}
\alias{get_small_index}
\title{Converts big index to range of small indexes}
\usage{
get_small_index(n)
}
\arguments{
\item{n}{big index (integer)}
}
\value{
range of small indexes
}
\description{
Converts big index to range of small indexes
}
