% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_class.R
\name{is.sudoku}
\alias{is.sudoku}
\title{Sudoku class utils
Function to check if an object is sudoku class.}
\usage{
is.sudoku(x)
}
\arguments{
\item{x}{an R object}
}
\value{
boolean flag
}
\description{
Sudoku class utils
Function to check if an object is sudoku class.
}
