% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_class.R
\name{print.sudoku}
\alias{print.sudoku}
\title{Method for printing sudoku object in predefinied format and coulors}
\usage{
\method{print}{sudoku}(mat, ...)
}
\arguments{
\item{mat}{matrix or sudoku class object}

\item{...}{ignored}
}
\value{
printed sudoku matrix in cat()
}
\description{
Method for printing sudoku object in predefinied format and coulors
}
