% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku_class.R
\name{print_matrix_cell}
\alias{print_matrix_cell}
\title{Print matrix cell with appropriate colour}
\usage{
print_matrix_cell(mat, check, row, col)
}
\arguments{
\item{mat}{matrix of sudoku class object}

\item{check}{list with checks performed on matrix}

\item{row}{row index}

\item{col}{column index}
}
\description{
Print matrix cell with appropriate colour
}
