% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{randomize_matrix}
\alias{randomize_matrix}
\title{Modify matrix by various approaches}
\usage{
randomize_matrix(mat, n = as.integer(runif(1, 50, 100)))
}
\arguments{
\item{mat}{sudoku matrix}

\item{n}{how many modification steps will be performed}
}
\value{
modified matrix
}
\description{
Modify matrix by various approaches
}
