% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{replace}
\alias{replace}
\title{Changes two values by each other}
\usage{
replace(mat, val1, val2)
}
\arguments{
\item{mat}{matrix or sudoku class object}

\item{val1}{value_1 to be changed by value_2}

\item{val2}{value_2 to be changed by value_1}
}
\value{
modified matrix
}
\description{
Changes two values by each other
}
