% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rotate}
\alias{rotate}
\title{Rotate matrix by 90 degrees}
\usage{
rotate(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
modified matrix
}
\description{
Rotate matrix by 90 degrees
}
