% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shift_index}
\alias{shift_index}
\title{Move elements from tail to the head of vector}
\usage{
shift_index(x, n)
}
\arguments{
\item{x}{vecor to be shifted}

\item{n}{shift value}
}
\value{
shifted vector
}
\description{
Move elements from tail to the head of vector
}
