% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shuffle}
\alias{shuffle}
\title{Internal shuffle function}
\usage{
shuffle(mat, if_row = TRUE)
}
\arguments{
\item{mat}{matrix or sudoku class object}

\item{if_row}{boolean flag indicating if suffle rows or columns}
}
\description{
Internal shuffle function
}
