% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shuffle_columns}
\alias{shuffle_columns}
\title{Change columns order randomly}
\usage{
shuffle_columns(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
modified matrix
}
\description{
Change columns order randomly
}
