% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shuffle_rows}
\alias{shuffle_rows}
\title{Change rows order randomly}
\usage{
shuffle_rows(mat)
}
\arguments{
\item{mat}{matrix or sudoku class object}
}
\value{
modified matrix
}
\description{
Change rows order randomly
}
