% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{start_game}
\alias{start_game}
\title{Main function to start game}
\usage{
start_game(difficulty = "E")
}
\arguments{
\item{difficulty}{Game difficulty level. Three levels allowed (E)asy, (M)edium, (H)ard.}
}
\value{
None
}
\description{
Generate sudoku matrix and open an interactive game session.
}
